+++ CHANGELOG.md
@@ -8,6 +8,15 @@ Updates should follow the [Keep a CHANGELOG](http://keepchangelog.com/) principl
 
 Lib part is missing TDR calculation and group correlation analysis.
 
+## v0.2.3 @avrbanac - corrected app graphs
+
+### Added
+- graph coloring
+
+### Fixed
+- switched from barchart to stacked barchart so that dB tap graph can be shown in negative values all the way to the zero for the main tap
+- fixed broken javafx stacked barchart so that it can correctly show negative values (explained in the FixedStackedBarChart class)
+
 ## v0.2.2 @avrbanac - not a module anymore, app added
 
 ### Added